% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asum4.R
\docType{data}
\name{asum4}
\alias{asum4}
\title{Combined alpha diversity data for meta-analysis.}
\format{
A dataframe with 16 rows and 18 variables.
}
\source{
\href{https://gordonlab.wustl.edu/supplemental-data/supplemental-data-portal/subramanian-et-al-2014/}{Gordon Lab}
}
\usage{
data(asum4)
}
\description{
Result outputs of differential analysis of alpha diversity
using linear or linear mixed effects model
from "alpha.compare" function combined from
4 studies for meta-analysis. The comparison was between gender
adjusted for age of infants at sample collection.
}
\examples{
data(asum4)
}
\references{
Subramanian et al. Nature. 2014 Jun 19; 510(7505): 417–421.
(\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4189846/}{PubMed})

Bender et al. Sci Transl Med. 2016 Jul 27; 8(349): 349ra100.
(\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5301310/}{PubMed})

Pannaraj et al. JAMA Pediatr. 2017;90095(7):647–54.
(\href{https://pubmed.ncbi.nlm.nih.gov/28492938/}{PubMed})

Thompson et al. Front Cell Infect Microbiol. 2015;5:3.
(\href{https://pubmed.ncbi.nlm.nih.gov/25705611/}{PubMed})
}
\keyword{datasets}
