% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtab.3stud.R
\docType{data}
\name{gtab.3stud}
\alias{gtab.3stud}
\title{Test datasets for microbiome age prediction.}
\format{A list of three dataframes.}
\source{
\href{https://gordonlab.wustl.edu/Subramanian_6_14/Nature_2014_Processed_16S_rRNA_datasets.html}{Gordon Lab}
}
\usage{
data(gtab.3stud)
}
\description{
A list of three datasets to be used as test datasets
for microbiome age prediction using Random Forest model.
}
\examples{
data(gtab.3stud)
}
\references{
Subramanian et al. Nature. 2014 Jun 19; 510(7505): 417–421.
(\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4189846/}{PubMed})

Bender et al. Sci Transl Med. 2016 Jul 27; 8(349): 349ra100.
(\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5301310/}{PubMed})

Pannaraj et al. JAMA Pediatr. 2017;90095(7):647–54.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/28492938}{PubMed})

Thompson et al. Front Cell Infect Microbiol. 2015;5:3.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/25705611?dopt=Abstract}{PubMed})
}
\keyword{datasets}
\keyword{test}
