% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.compare.R
\name{taxa.compare}
\alias{taxa.compare}
\title{Compare taxa relative abundance}
\usage{
taxa.compare(taxtab, propmed.rel = "gamlss", comvar, adjustvar,
  personid = "personid", longitudinal = "yes", percent.filter = 0.05,
  relabund.filter = 5e-05, p.adjust.method = "fdr", ...)
}
\arguments{
\item{taxtab}{taxa relative abundance table (already merged to mapping file) from phylum to species or any preferred highest taxa level.}

\item{propmed.rel}{statistical method for comparing relative abundance. Options are "lm" for LM/LMEM or "gamlss" for GAMLSS with BEZI family.}

\item{comvar}{main variable for comparison}

\item{adjustvar}{variables to be adjusted.}

\item{personid}{name of variable for person id (applicable for longitudinal data)}

\item{longitudinal}{whether data is longitudinal? Options are "yes" or "no". Default is "yes".}

\item{percent.filter}{prevalence threshold (the percentage of number of samples the taxa/pathway available). Default is 0.05.}

\item{relabund.filter}{relative abundance threshold (the minimum of the average relative abundance for a taxa/pathway to be retained). Default is 0.00005.}

\item{p.adjust.method}{method for multiple testing adjustment. Options are those of the p.adjust.methods of stats:: p.adjust function. Default for this function is "fdr".}

\item{pooldata}{whether data is pooled from multiple studies. Default is FALSE.}
}
\value{
matrice of coefficients, standard errors, p-values and multiple testing adjusted p-values of all variables in the models.
}
\description{
This function compares taxa relative abundance summary tables at all levels between groups using GAMLSS with BEZI or Linear/Linear Mixed Effect models (LM/LMEM) after filtering (using prevalence and relative abundance thresholds).
}
\examples{
#Load summary tables of bacterial taxa relative abundance from Bangladesh data
data(taxtab.rm7)
#Comparison of bacterial taxa relative abundance using LMEM or GAMLSS (take time to run)
taxacom6.rmg<-taxa.compare(taxtab=taxtab6.rm[[5]],taxsum="rel",propmed.rel="lm",comvar="bf",adjustvar="age.sample",longitudinal="yes",p.adjust.method="fdr")
taxacom6.zi.rmg<-taxa.compare(taxtab=taxtab6.rm[[5]],taxsum="rel",propmed.rel="gamlss",comvar="bf",adjustvar="age.sample",longitudinal="yes",p.adjust.method="fdr")
taxcomtab.show(taxcomtab=taxacom6.zi.rmg,tax.select="none", showvar="bfNon_exclusiveBF", tax.lev="l2",readjust.p=TRUE,p.adjust.method="fdr")
}
\keyword{abundance}
\keyword{comparison}
\keyword{taxa}
