% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.niceplot.R
\name{meta.niceplot}
\alias{meta.niceplot}
\title{Nice meta-analysis plots.}
\usage{
meta.niceplot(metadat, sumtype = "taxa", level = "main", p = "p.nebf",
  p.adjust = "p.adjust.nebf", phyla.col = c("select", "rainbow"),
  leg.key.size = 1, leg.text.size = 8, heat.text.x.size = 8,
  heat.text.x.angle = 0, forest.axis.text.y = 8, forest.axis.text.x = 8)
}
\arguments{
\item{metadat}{output data from metatab.show.}

\item{sumtype}{Either "taxa" for taxa and "path" for pathway.}

\item{level}{"main" for main level such as phylum or "sub" for higher level such as species. Default is "main".}

\item{p}{name of variable for p-values}

\item{p.adjust}{name of variable for multiple testing adjusted p-values}

\item{phyla.col}{type of color for main level (phylum). Options are "select" (default) or "rainbow".}

\item{leg.key.size}{legdend key size for heatmap.}

\item{leg.text.size}{legend text size for heatmap.}

\item{heat.text.x.size}{heatmap x label text size.}

\item{heat.text.x.angle}{heatmap x label text angle.}

\item{forest.axis.text.y}{forest plot y label text size.}

\item{forest.axis.text.x}{forest plot x label text size.}
}
\value{
combined heatmap forest plot.
}
\description{
This function displays meta-analysis results of relative abundance as a nice combined heatmap and forest plot.
}
\examples{
#Load saved results of four studies for the comparison of bacterial taxa relative abundance between genders adjusted for breastfeeding and infant age at sample collection
data(taxacom.rm.sex.adjustbfage)
data(taxacom.ha.sex.adjustbfage)
data(taxacom6.zi.usbmk.sex.adjustbfage)
data(taxacom6.unc.sex.adjustedbfage)
taxacom6.zi.rm.sex.adjustbfage$study<-"Subramanian et al 2014 (Bangladesh)"
taxacom6.zi.rm.sex.adjustbfage$pop<-"Bangladesh"
taxacom.zi.ha.sex.adjustbfage$study<-"Bender et al 2016 (Haiti)"
taxacom.zi.ha.sex.adjustbfage$pop<-"Haiti"
taxacom6.zi.usbmk.sex.adjustbfage$study<-"Pannaraj et al 2017 (USA(CA_FL))"
taxacom6.zi.usbmk.sex.adjustbfage$pop<-"USA(CA_FL)"
taxacom6.zi.unc.sex.adjustedbfage$study<-"Thompson et al 2015 (USA(NC))"
taxacom6.zi.unc.sex.adjustedbfage$pop<-"USA(NC)"
tabsex4<-plyr::rbind.fill(taxacom6.zi.rm.sex.adjustbfage,taxacom.zi.ha.sex.adjustbfage,taxacom6.zi.usbmk.sex.adjustbfage,taxacom6.zi.unc.sex.adjustedbfage)
#Meta-analysis (take time to run)
metab.sex<-meta.taxa(taxcomdat=tabsex4,summary.measure="RR",pool.var="id",studylab="study",backtransform=FALSE,percent.meta=0.5,p.adjust.method="fdr")
#nice plot phylum level
metadat<-metatab.show(metatab=metab.sex$random,com.pooled.tab=tabsex4,tax.lev="l2",showvar="genderMale",p.cutoff.type="p", p.cutoff=1,display="data")
meta.niceplot(metadat=metadat,sumtype="taxa",level="main",p="p",p.adjust="p.adjust",phyla.col="rainbow",leg.key.size=1,leg.text.size=8,heat.text.x.size=7,heat.text.x.angle=0,forest.axis.text.y=8,forest.axis.text.x=7)
#nice plot family level
metadat<-metatab.show(metatab=metab.sex$random,com.pooled.tab=tabsex4,tax.lev="l5",showvar="genderMale",p.cutoff.type="p", p.cutoff=1,display="data")
meta.niceplot(metadat=metadat,sumtype="taxa",level="sub",p="p",p.adjust="p.adjust",phyla.col="rainbow",leg.key.size=1,leg.text.size=8,heat.text.x.size=7,forest.axis.text.y=8,forest.axis.text.x=7)
}
\keyword{forest}
\keyword{heatmap}
\keyword{meta-analysis}
\keyword{plot.}
