% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.taxa.R
\name{meta.taxa}
\alias{meta.taxa}
\title{Meta-analysis of taxa/pathway abundance comparison.}
\usage{
meta.taxa(taxcomdat, estimate.pattern = "Estimate.",
  se.pattern = "Std. Error.", summary.measure = "RR",
  pool.var = "id", studylab = "study", backtransform = FALSE,
  percent.meta = 0.5, p.adjust.method = "fdr", ...)
}
\arguments{
\item{taxcomdat}{matrice of estimates and SE of all taxa/pathways combined from all included studies.}

\item{estimate.pattern}{string pattern for estimates. Default is "Estimate.".}

\item{se.pattern}{string pattern for standard error. Default is "Std. Error.".}

\item{summary.measure}{"RR" for estimates from GAMLSS with BEZI family and "RD" for estimates from Linear/linear mixed effect model. Default is "RR"}

\item{pool.var}{name of id variable for meta-analysis. Default is "id".}

\item{studylab}{name of variable characterizing included studies. Default is "study".}

\item{backtransform}{whether or not to perform backtransformation of the estimates. Default is FALSE.}

\item{percent.meta}{the threshold percentage of number of studies that a taxa is available to do meta-analysis. Default is 0.5}

\item{p.adjust.method}{method for multiple testing adjustment (available methods of the function p.adjust). Default is "fdr".}
}
\value{
a list of matrices of results for all variables in the comparison models.
}
\description{
This function does meta-analysis based on estimates and standard errors from taxa/pathway abundance comparison using random effect and fixed effect meta-analysis models.
}
\examples{
#Load saved results of four studies for the comparison of bacterial taxa relative abundance between genders adjusted for breastfeeding and infant age at sample collection
data(taxacom.rm.sex.adjustbfage)
data(taxacom.ha.sex.adjustbfage)
data(taxacom6.zi.usbmk.sex.adjustbfage.)
data(taxacom6.unc.sex.adjustedbfage)
taxacom6.zi.rm.sex.adjustbfage$study<-"Subramanian et al 2014 (Bangladesh)"
taxacom6.zi.rm.sex.adjustbfage$pop<-"Bangladesh"
taxacom.zi.ha.sex.adjustbfage$study<-"Bender et al 2016 (Haiti)"
taxacom.zi.ha.sex.adjustbfage$pop<-"Haiti"
taxacom6.zi.usbmk.sex.adjustbfage$study<-"Pannaraj et al 2017 (USA(CA_FL))"
taxacom6.zi.usbmk.sex.adjustbfage$pop<-"USA(CA_FL)"
taxacom6.zi.unc.sex.adjustedbfage$study<-"Thompson et al 2015 (USA(NC))"
taxacom6.zi.unc.sex.adjustedbfage$pop<-"USA(NC)"
tabsex4<-plyr::rbind.fill(taxacom6.zi.rm.sex.adjustbfage,taxacom.zi.ha.sex.adjustbfage,taxacom6.zi.usbmk.sex.adjustbfage,taxacom6.zi.unc.sex.adjustedbfage)
#Meta-analysis (take time to run)
metab.sex<-meta.taxa(taxcomdat=tabsex4,summary.measure="RR",pool.var="id",studylab="study",backtransform=FALSE,percent.meta=0.5,p.adjust.method="fdr")
metatab.show(metatab=metab.sex$random,com.pooled.tab=tabsex4,tax.lev="l2",showvar="genderMale",p.cutoff.type="p", p.cutoff=0.2,display="table")
}
\keyword{abundance}
\keyword{meta-analysis.}
