% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metatab.show.R
\name{metatab.show}
\alias{metatab.show}
\title{Display meta-analysis results.}
\usage{
metatab.show(metatab, com.pooled.tab, sumvar = "taxa", tax.lev = "l2",
  showvar, estimate.pattern = "Estimate.", se.pattern = "Std. Error.",
  p.pattern = "Pr(>|t|)", readjust.p = FALSE, p.cutoff.type = "p",
  p.cutoff = 0.05, display = "plot", plot = "heatmap",
  fill.value = "log(OR)", grid = FALSE, digit = 2, p.digit = 4,
  ...)
}
\arguments{
\item{metatab}{matrice of taxa/pathway abundance comparison meta-analysis results generated from meta.taxa.}

\item{com.pooled.tab}{matrice of taxa/pathway abundance comparison generated from taxa.compare or pathway.compare combined from all included studies.}

\item{sumvar}{Either "taxa" for taxa and "path" for pathway.}

\item{tax.lev}{taxa level to be displayed. Options are from "l2" (phylum) to "l7" (species). Default is "l2".}

\item{showvar}{variable (string pattern) in the model to be displayed.}

\item{estimate.pattern}{string pattern for estimates. Default is "Estimate.".}

\item{se.pattern}{string pattern for standard error. Default is "Std. Error.".}

\item{readjust.p}{multiple testing re-adjustment for only the level to be displayed (TRUE) or keep original multiple testing adjustment for all taxa of all levels (FALSE).Default is FALSE.}

\item{p.cutoff.type}{type of p-value for cutoff. Options are "p" for p-value or "p.adjust" for multiple testing adjusted p-value. Default is "p".}

\item{p.cutoff}{cutoff p-value to be displayed. Default is 0.05.}

\item{display}{type of display. Options are display=c("plot","table","data")}

\item{plot}{type of plot. Options are plot=c("heatmap","forest").}

\item{fill.value}{name of legend.}

\item{grid}{whether multiple plots will be displayed alongside. Default is FALSE.}

\item{digit}{digit for estimates and 95 CI. Default is 2.}

\item{p.digit}{digit for p-values. Default is 4.}
}
\value{
plot table or data.
}
\description{
This function displays meta-analysis results of relative abundance as heatmap, forest plot, table or data.
}
\examples{
#Load saved results of four studies for the comparison of bacterial taxa relative abundance between genders adjusted for breastfeeding and infant age at sample collection
data(taxacom.rm.sex.adjustbfage)
data(taxacom.ha.sex.adjustbfage)
data(taxacom6.zi.usbmk.sex.adjustbfage.)
data(taxacom6.unc.sex.adjustedbfage)
taxacom6.zi.rm.sex.adjustbfage$study<-"Subramanian et al 2014 (Bangladesh)"
taxacom6.zi.rm.sex.adjustbfage$pop<-"Bangladesh"
taxacom.zi.ha.sex.adjustbfage$study<-"Bender et al 2016 (Haiti)"
taxacom.zi.ha.sex.adjustbfage$pop<-"Haiti"
taxacom6.zi.usbmk.sex.adjustbfage$study<-"Pannaraj et al 2017 (USA(CA_FL))"
taxacom6.zi.usbmk.sex.adjustbfage$pop<-"USA(CA_FL)"
taxacom6.zi.unc.sex.adjustedbfage$study<-"Thompson et al 2015 (USA(NC))"
taxacom6.zi.unc.sex.adjustedbfage$pop<-"USA(NC)"
tabsex4<-plyr::rbind.fill(taxacom6.zi.rm.sex.adjustbfage,taxacom.zi.ha.sex.adjustbfage,taxacom6.zi.usbmk.sex.adjustbfage,taxacom6.zi.unc.sex.adjustedbfage)
#Meta-analysis (take time to run)
metab.sex<-meta.taxa(taxcomdat=tabsex4,summary.measure="RR",pool.var="id",studylab="study",backtransform=FALSE,percent.meta=0.5,p.adjust.method="fdr")
metatab.show(metatab=metab.sex$random,com.pooled.tab=tabsex4,tax.lev="l2",showvar="genderMale",p.cutoff.type="p", p.cutoff=0.2,display="table")
metatab.show(metatab=metab.sex$random,com.pooled.tab=tabsex4,tax.lev="l2",showvar="genderMale",p.cutoff.type="p", p.cutoff=1,display="plot",plot="heatmap")
}
\keyword{abundance}
\keyword{display.}
\keyword{meta-analysis}
