% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomeage.R
\name{microbiomeage}
\alias{microbiomeage}
\title{Predict microbiome age.}
\usage{
microbiomeage(l6.relabundtab)
}
\arguments{
\item{l6.relabundtab}{list of taxa summary table from phylum up to genus level merged to mapping file outputed from QIIME of all included studies.}
}
\value{
list of training and test sets of Bangladesh data, shared genera list, relative abundance data of shared genera, randomforest fit, RF model performance plot,predicted microbiome age of Bangladesh data and data of other included studies.
}
\description{
This function predicts microbiome age using Random Forest model based on relative abundances of bacterial genera shared with the Bangladesh study (Subramanian et al 2014).
This function gets the shared genera list between the Bangladesh study and all other included studies,
get the training and test sets from Bangladesh data based on the shared genera list,
fit the train Random Forest model and predict microbiome age in the test set of Bangladesh data and data from all included studies,
check for performance of the model based on the shared genera list on Bangladesh healthy cohort data, reproduce the findings of the Bangladesh malnutrition study.
}
\examples{
#Load data from each study and put in a list
bal6<-read.delim(system.file("extdata/QIIME_outputs/Bangladesh/tax_mapping7", "Subramanian_et_al_mapping_file_L6.txt", package = "metamicrobiomeR", mustWork = TRUE))
colnames(bal6)<-tolower(colnames(bal6))
uncl6<-read.delim(system.file("extdata/QIIME_outputs/USA(NC)/tax_mapping7", "mapping_fileall_L6.txt", package = "metamicrobiomeR", mustWork = TRUE))
colnames(uncl6)<-tolower(colnames(uncl6))
uncl6$age.sample<-uncl6$agemo
usl6<-read.delim(system.file("extdata/QIIME_outputs/USA(CA_FL)/tax_mapping7", "USBMK_Mapping_122414.filtered.with_pairing_status_L6.txt", package = "metamicrobiomeR", mustWork = TRUE))
colnames(usl6)<-tolower(colnames(usl6))
usl6$age.sample<-usl6$bbage/30
hal6<-read.delim(system.file("extdata/QIIME_outputs/Haiti/tax_mapping7", "Haiti_Mapping.complete_L6.txt", package = "metamicrobiomeR", mustWork = TRUE))
colnames(hal6)<-tolower(colnames(hal6))
hal6$age.sample<-hal6$bbage/30
gtab<-list(unc=uncl6,ca_fl=usl6[usl6$sampletype=="STL"&usl6$mombb=="Baby",],haiti=hal6[hal6$sampletype=="STL"&hal6$mombb=="B",])
#predict microbiome age (take time to run)
miage<-microbiomeage(l6.relabundtab=gtab)
#list of shared genera that are available in the Bangladesh study and other included studies
kable(miage$sharedgenera.importance)
#check performance
grid.arrange(miage$performanceplot$ptrain, miage$performanceplot$ptest,nrow=1)
#replicate the findings of Subramanian et al paper
ggplot2::ggplot() +geom_point(data=miage$microbiomeage.bangladesh$all,aes(x=age.sample, y=age.predicted, colour=health_analysis_groups))
}
\keyword{Forest.}
\keyword{Random}
\keyword{age}
\keyword{microbiome}
