% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.compare.R
\name{pathway.compare}
\alias{pathway.compare}
\title{Compare (kegg) pathway abundance}
\usage{
pathway.compare(pathtab, mapfile, sampleid = "sampleid", pathsum = "rel",
  stat.med = "gamlss", transform = "none", comvar, adjustvar,
  personid = "personid", longitudinal = "yes", p.adjust.method = "fdr",
  percent.filter = 0.05, relabund.filter = 5e-05, pooldata = FALSE,
  age.limit = 1e+06, age.lowerlimit = 0, ...)
}
\arguments{
\item{pathtab}{list of pathway abundance table of all levels.}

\item{mapfile}{mapping file or file containing covariates.}

\item{sampleid}{variable containing sample id to be matched between pathway abundance table and mapping file.}

\item{pathsum}{type of abundance to be compared. Options are "rel" for relative abundance or "log" for log of absolute abundance. Default is "rel".}

\item{stat.med}{statistical method for comparison. stat.med can be "lm" for LM/LMEM (usable for both pathsum="rel" or "log") or "gamlss" for GAMLSS with BEZI family (gamlss only make sense if pathsum="rel" ).}

\item{comvar}{main variable for comparison.}

\item{adjustvar}{variables to be adjusted.}

\item{personid}{name of variable for person id in mapping file (applicable for longitudinal data)}

\item{longitudinal}{whether the data is longitudinal. Default is "yes".}

\item{p.adjust.method}{method for multiple testing adjustment. Available options are those of the p.adjust function. Default is "fdr".}

\item{percent.filter}{prevalence threshold (the percentage of number of samples the taxa/pathway available). Default is 0.05.}

\item{relabund.filter}{relative abundance threshold (the minimum of the average relative abundance for a taxa/pathway to be retained). Default is 0.00005.}

\item{pooldata}{whether the data is pooled from multiple studies. Default is FALSE.}

\item{age.limit}{upper age limit for data to be included. Default is 1000000 months (~no upper age limit).}

\item{age.lowerlimit}{lower age limit for data to be included. Default is 0 month.}
}
\value{
matrice of coefficients, standard errors, p-values and multiple testing adjusted p-values of all variables in the models.
}
\description{
This is a slightly modified version of the taxa.compare function.
It compares pathway abundance generated from PICRUSt analysis between groups using different methods (apply to pathway summary tables already merged to mapping file and put in a list (e.g.level 1, 2 and 3)).
Specifically, it compares relative abundances of bacterial functional pathways at all levels using GAMLSS or LM/LMEM and compares of log(absolute abundances) of bacterial functional pathways at all levels using LM/LMEM.
}
\examples{
#Load Bangladesh extra metadata
data(sam.rm)
#Read in PICRUSt output for KEGG pathways level 1-3
patht<-system.file("extdata/QIIME_outputs/Bangladesh/picrust", package = "metamicrobiomeR", mustWork = TRUE)
kegg<-read.multi(patht=patht,patternt=".txt",assignt="no")
kegg.rm<-list()
for (i in 1:length(kegg)){
  rownames(kegg[[i]])<-kegg[[i]][,"kegg_pathways"]
  kegg[[i]]<-kegg[[i]][,colnames(kegg[[i]])[!colnames(kegg[[i]]) \%in\% c("otu.id","kegg_pathways")]]
  kegg.rm[[i]]<-as.data.frame(t(kegg[[i]]))
}
covar.rm<-merge(samde, he50[,c("child.id","gender","zygosity","day.firstsample","day.lastsample","n.sample","sampling.interval.msd","month.exbf","month.food",
                               "n.diarrhea.yr","percent.time.diarrhea","fraction.antibiotic","subject.allocation")], by="child.id")
covar.rm<-dplyr::rename(covar.rm,sampleid=fecal.sample.id, personid=child.id ,age.sample=age.months)
covar.rm$bf<-factor(covar.rm$bf, levels=c('ExclusiveBF','Non_exclusiveBF','No_BF'))
covar.rm$personid<-as.factor(covar.rm$personid)
#Comparison of pathway relative abundances (take time to run)
pathcom.rm6.rel.gamlss.sexg<-pathway.compare(pathtab=kegg.rm,mapfile=covar.rm,sampleid="sampleid",pathsum="rel",stat.med="gamlss",comvar="gender",adjustvar=c("age.sample","bf"),longitudinal="yes",p.adjust.method="fdr",percent.filter=0.05,relabund.filter=0.00005,age.limit=6)
taxcomtab.show(taxcomtab=pathcom.rm6.rel.gamlss.sexg$l3, sumvar="path",tax.lev="l3",tax.select="none",showvar="genderMale", p.adjust.method="fdr",p.cutoff=0.05,digit=4)
}
\keyword{abundance}
\keyword{comparison.}
\keyword{pathway}
