% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.multi.R
\name{read.multi}
\alias{read.multi}
\title{Read multiple files}
\usage{
read.multi(patht, patternt = ".txt", assignt = "no", study = NULL)
}
\arguments{
\item{patht}{path to files.}

\item{patternt}{pattern of files. Options are ".txt" or ".csv".}

\item{assignt}{assign files. Default is "no".}

\item{study}{name of the study. Default is NULL.}
}
\value{
list of all data files in the path
}
\description{
This function reads multiple files of the same pattern in a directory into R.
}
\examples{
patht<-system.file("extdata/QIIME_outputs/Bangladesh/alpha_div_collated", package = "metamicrobiomeR", mustWork = TRUE)
alpha.ba<-read.multi(patht=patht,patternt=".txt",assignt="no",study="Bangladesh")
}
\keyword{files}
\keyword{multiple}
